refactor(vscode): lift initial preview-card DOM build into cardBuilder.ts
Stepping stone toward the eventual `<preview-card>` Lit component:
extracts `behavior.ts`'s `createCard` function (~150 lines of imperative
DOM build) into a dedicated typed module. The dynamic update paths
(`updateCardMetadata`, `updateImage`, `applyA11yUpdate`) still live in
`behavior.ts` and patch the card in place via `document.getElementById`
— that's the next slice once `<preview-card>` lands and the two paths
can fold into a single reactive `render()`.

New `cardBuilder.ts` exports:
- `buildPreviewCard(p: PreviewInfo, config: CardBuilderConfig)
   → HTMLElement` — pure-DOM builder, returns a detached `<div>` ready
  for the caller to insert.
- `CardBuilderConfig` — typed dependency surface
  (`vscode`, `cardCaptures` map, `staleBadge`, `frameCarousel`,
  `liveState`, `earlyFeatures()`, `inFocus()`, `enterFocus(card)`,
  `exitFocus()`, `observeForViewport(card)`). Captures every
  collaborator the imperative function previously closed over so the
  builder doesn't need closure access to `setupPreviewBehavior`.

`behavior.ts`'s `createCard` is now a one-liner that calls
`buildPreviewCard(p, cardBuilderConfig)` with the config built once at
panel boot. No behaviour change — same DOM shape, same event handlers,
same side effects (`cardCaptures.set`, `staleBadge.apply(card, false)`,
`observeForViewport(card)`).

`behavior.ts`: 1273 → 1141 lines (−132 LOC).

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 350/350 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke: open the panel, verify per-card affordances:
  - Title click opens the source file (`openFile`).
  - Focus button toggles into / out of focus mode.
  - Animation icon shows in the title row for animated previews.
  - Image-container click enters LIVE for non-stale, non-already-live
    cards; respects Shift+click multi-stream semantics.
  - Stale-badge button refreshes a stale card.
  - A11y legend + overlay layer are attached when `earlyFeatures` is
    on and `p.a11yFindings.length > 0`.
  - Frame carousel shows below animated cards.
  - Card is observed by the viewport tracker (lazy thumbs / scroll-
    ahead daemon hint).
